fix(utils): move file to utils package
