+++ packages/did-provider-jwk/src/types/jwk-provider-types.ts
@@ -1,3 +1,5 @@
+import { JwkDidSupportedKeyTypes, KeyUse } from '@veramo/utils'
+
 export type JwkCreateIdentifierOptions = {
   keyType?: JwkDidSupportedKeyTypes
   privateKeyHex?: string
@@ -13,14 +15,3 @@ type ImportOrGenerateKeyOpts = {
   keyType: JwkDidSupportedKeyTypes
   privateKeyHex?: string
 }
-
-export enum SupportedKeyTypes {
-  Secp256r1 = 'Secp256r1',
-  Secp256k1 = 'Secp256k1',
-  Ed25519 = 'Ed25519',
-  X25519 = 'X25519',
-}
-
-export type JwkDidSupportedKeyTypes = keyof typeof SupportedKeyTypes
-
-export type KeyUse = 'sig' | 'enc'
