+++ packages/utils/src/index.ts
@@ -9,3 +9,4 @@ export * from './credential-utils.js'
 export * from './did-utils.js'
 export * from './encodings.js'
 export * from './type-utils.js'
+export * from './jwk-did-utils.js'
